Let RIOHandler handle R outputs which are absolute file paths
RIOHandler cannot handle outputs with an absolute file which reside outside
the actual working directory. However, this may a valid use case as R scripts
should be free to either store outputs within a local directory or completely
elsewhere in the system.

The full file handle created by the R output result string can  be used instead
of a file handle based on the basename. However, as this is true at least for
GenericFileDataBinding or GenericFileDataWithGTBinding types. Other types may
require the file handle based on basename.

Resolves: #274