Ensure a proper layerName is being put in result URL
The WMS uploader have to use the layerName (formerly named
storeName) as name for the uploading coverage. Otherwise
the WMS layerNames parameter value for the resulting GetMap
output will point to a non-existing layer.

Making a shorter unique layer name is now also being applied
to WCS-/WFS-Generator, as long random layernames are hard to
associate with.